Automata4j-core: refactored transitions descendants to transition package
